requirements.txt
annotated-types==0.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama @ file:///C:/Users/dev-admin/buildout/croot/colorama_1759952627215/work
filelock==3.24.3
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049098810/work
pillow==12.1.1
platformdirs==4.9.2
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
requests==2.32.5
sentry-sdk==2.53.0
setuptools==80.10.2
smmap==5.0.2
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
tqdm @ file:///C:/Users/task_177139866619957/croot/tqdm_1771398766117/work
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wandb==0.25.0
wheel==0.46.3
